Updated server-side files
Please note, these files will not work with Firebase. Also, the requests to the data store don't actually work (oops). I've figured that the issue is the permissions/authentication, which needed to be changed anyway so we can keep the storage bucket private. At the time of this commit, honestly just too tired to keep working on it.