Changed implementaton of dicomweb commands-untested and unimplemented
DICOM web commands have begun being implemented
I made a bit of progress on two different implementations--through nodejs and python. The python approach will be used if we ultimately feel a need to switch to flask, however it presents issues I'm not familiar dealing with, while I am familiar with the nodejs issues.