Implementation of store command (#9)
* Updated client-side files

Should be mostly done as far as this branch is concerned--wanted to enable uploading a file to the server, but that is all that's left to be implemented for this task

* Updated server-side files

Please note, these files will not work with Firebase. Also, the requests to the data store don't actually work (oops). I've figured that the issue is the permissions/authentication, which needed to be changed anyway so we can keep the storage bucket private. At the time of this commit, honestly just too tired to keep working on it.

* Fixed authentication(?)

Yeah, need to upload the env file but that should give the webapp authentication

* Create .env

* Changed implementaton of dicomweb commands-untested and unimplemented

DICOM web commands have begun being implemented
I made a bit of progress on two different implementations--through nodejs and python. The python approach will be used if we ultimately feel a need to switch to flask, however it presents issues I'm not familiar dealing with, while I am familiar with the nodejs issues.

* Updating handleDicom to reflect dicomweb commands

* Updating web files

Implemented some things to allow for testing/implementation of the store command

* Implemented Dicomweb store