Completed Study Search (#10)
* Prepping implementation of Searching

finished with backend implementation(before testing) working on front end implementation

* Updating? part 1

* Updating? Part 2

* Updating server file

* Finished parsing study stuff

Now just need to implement them in a search bar, which I've begun implementing, but I thought this was a good checkpoint

* Updated UI--just styling left 

This is the first search bar, for searching through the studies and allowing selection. I've not yet linked it up to retrieve, though I've written a place for it.

* Updating server side

Honestly can't recall if I made any major changes here, primarily just updating the README

* I believe I'm done for now with the study search

In the future may work on the U/I a bit more
Still need to implement the search for the series held locally, but at the moment I don't know what the local storage situation is going to look like well enough to get a significant start on it.